+++ app_version.py
@@ -1,4 +1,4 @@
 """Canonical application/release version constants."""
 
-RELEASE_VERSION = "v2.13"
+RELEASE_VERSION = "v2.15"